Conflicts:
	app/src/main/java/deneme/example/filedeneme/Main3Activity.kt
	app/src/main/java/deneme/example/filedeneme/MainActivity.kt
	app/src/main/java/deneme/example/filedeneme/RecyclerViewAdapter.kt
	app/src/main/res/layout/recycler_view_item.xml
	com.mobilex.loggerbird/build.gradle
	com.mobilex.loggerbird/src/main/AndroidManifest.xml
	com.mobilex.loggerbird/src/main/java/fragments/FragmentLoggerBird.kt
	com.mobilex.loggerbird/src/main/java/observers/LogActivityLifeCycleObserver.kt
	com.mobilex.loggerbird/src/main/java/paint/PaintActivity.kt
	com.mobilex.loggerbird/src/main/java/paint/PaintFabAnimation.kt
	com.mobilex.loggerbird/src/main/java/paint/PaintView.kt
	com.mobilex.loggerbird/src/main/res/layout/activity_paint.xml
	com.mobilex.loggerbird/src/main/res/layout/activity_paint_save_snackbar.xml
	com.mobilex.loggerbird/src/main/res/layout/fragment_logger_bird.xml
	com.mobilex.loggerbird/src/main/res/values/colors.xml

+++ com.mobilex.loggerbird/src/main/java/utils/JiraAuthentication.kt
@@ -230,12 +230,14 @@ class JiraAuthentication {
     }
 
     @RequiresApi(Build.VERSION_CODES.JELLY_BEAN)
-    private fun jiraExceptionHandler(
+   internal fun jiraExceptionHandler(
         e: Exception? = null,
         filePathName: File? = null,
         throwable: Throwable? = null
     ) {
-        filePathName?.delete()
+        if(filePathName?.name != "logger_bird_details_old_session.txt"){
+            filePathName?.delete()
+        }
         LoggerBirdService.loggerBirdService.finishShareLayout("jira_error")
         e?.printStackTrace()
         LoggerBird.callEnqueue()
@@ -299,6 +301,7 @@ class JiraAuthentication {
                         )
                     } else {
                         activity.runOnUiThread {
+                            LoggerBirdService.loggerBirdService.detachProgressBar()
                             defaultToast.attachToast(
                                 activity = activity,
                                 toastMessage = context.resources.getString(R.string.unhandled_file_doesnt_exist)
